Split all methods into separated modules; fix
