+++ src/lib/markup.js
@@ -1,0 +1,8 @@
+export default `
+<div class="__PREFIX__box">
+  <div class="__PREFIX__content">
+    <div class="__PREFIX__icon"></div>
+    <div class="__PREFIX__text"></div>
+  </div>
+</div>
+`;
